v2-agent/requirements.txt
# Core FastAPI dependencies
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.3
pydantic-settings==2.2.1

# AWS SDK dependencies
boto3==1.29.3
botocore==1.32.3

# Environment and configuration
python-dotenv==1.0.1
pyyaml==6.0.1

# Lambda compatibility (for potential Lambda deployment)
mangum==0.17.0

# Testing dependencies
pytest==8.2.1
pytest-cov==5.0.0
hypothesis==6.100.0
